Implement OAuth2 login for Frontend
Added the "/login-by-code" API for the Next.js frontend to fetch credentials.
When a user logs in with a Google account, the JWT token and user data are stored in Redis and an auth code is generated.
By providing this auth code to the "/login-by-code" API, the frontend can retrieve the JWT and all relevant user session data required for login.